fe + detections and anomaly integrated

+++ detection/haliva-detection/detectors/anomaly_detection.py
@@ -10,7 +10,7 @@ import os
 
 tf.config.optimizer.set_experimental_options({'layout_optimizer': False})
 class AnomalyDetection:
-    def __init__(self, cache_dir="cache/", tolerance_frames=5):
+    def __init__(self, cache_dir="cache/", tolerance_frames=5, container_anomaly_process=None):
         model_file = "models/model_lstm_gil.hdf5"
         self.model = tf.keras.models.load_model(model_file, custom_objects={'LayerNormalization': LayerNormalization})
         self.batch_size = 3
@@ -21,6 +21,7 @@ class AnomalyDetection:
         self.tolerance_frames = tolerance_frames
         self.total_frames = 0
         self.done_tasks = [0]
+        self.container_anomaly_process = container_anomaly_process
 
     def __addToQueue(self):
         if self.queue.empty():
@@ -44,23 +45,40 @@ class AnomalyDetection:
             for j in range(0, 10):
                 clip[j] = test[i + j, :, :, :]
             sequences[i] = clip
+        print("anomaly: add to queue")
         next_cache_file: int = self.__addToQueue()
         np.save(self.cache_dir + str(next_cache_file), sequences)
         self.queue.put(next_cache_file)
 
+    def calc_processing_percents(self, video_capture):
+
+        # get next frame number out of all the frames for video
+        next_frame_no = video_capture.get(cv2.CAP_PROP_POS_FRAMES)
+        # get total number of frames in the video
+        total_frames = video_capture.get(cv2.CAP_PROP_FRAME_COUNT)
+
+        complete = round(next_frame_no/total_frames, 4)
+
+        self.container_anomaly_process.processing_percents = format(complete * 100, ".2f")
+
+        print("anomaly", self.container_anomaly_process.processing_percents, "%")
+
     def get_anomaly(self, video) -> np.array:
         def start_recording_thread():
+            print("anomaly: start_recording_thread")
             vs = cv2.VideoCapture(video)
             property_id = int(cv2.CAP_PROP_FRAME_COUNT)
             test_length = int(cv2.VideoCapture.get(vs, property_id))
-            self.total_frames = test_length
+            self.total_frames = vs.get(cv2.CAP_PROP_FRAME_COUNT)
             # test_length = len(list(filter(lambda x: '.jpg' in x, os.listdir(video[:-7]))))
             sz = self.sample_size if test_length >= self.sample_size else test_length
             test_length = test_length - sz
             test = np.zeros(shape=(sz, 256, 256, 1))
             cnt = 0
             while (1):
+                self.calc_processing_percents(vs)
                 if cnt == sz:
+                    print("anomaly: batch checkpoint")
                     self.get_single_test(test)
                     sz = self.sample_size if test_length >= self.sample_size else test_length
                     test_length = test_length - sz
@@ -68,6 +86,7 @@ class AnomalyDetection:
                     cnt = 0
                 ret, frame = vs.read()
                 if frame is None:
+                    print("anomaly: finish going over frames")
                     if test.any():
                         self.get_single_test(test)
                     break
@@ -82,29 +101,46 @@ class AnomalyDetection:
         self.rec_thread.start()
         srs = []
         video_counter = 0
-        while not self.queue.empty() or self.rec_thread.is_alive():
-            np_predict_array = str(self.__getFromQueue())
-            sequences = np.load(self.cache_dir + np_predict_array + ".npy")
-            os.remove(self.cache_dir + np_predict_array + ".npy")
-            sequences_shape = sequences.shape[0]
-            reconstructed_sequences = self.model.predict(sequences, batch_size=self.batch_size)
-            sequences_reconstruction_cost = np.array(
-                [np.linalg.norm(np.subtract(sequences[i], reconstructed_sequences[i])) for i in
-                 range(0, sequences_shape)])
-            sa: np.ndarray = (sequences_reconstruction_cost - np.min(sequences_reconstruction_cost)) / np.max(
-                sequences_reconstruction_cost)
-            srs.append((1.0 - sa, video_counter))
-            self.done_tasks[0] += 1
-            # plot the regularity scores
-            # plt.plot(1.0 - sa)
-            # plt.ylabel('regularity score Sr(t)')
-            # plt.xlabel('frame t')
-            # plt.show()
-            video_counter += 1
+        print("anomaly: before while not queue")
+        while self.rec_thread.is_alive():
+            #print("anomaly: inside thread is alive = true")
+            if not self.queue.empty():
+                print("anomaly: queue is not empty")
+                np_predict_array = str(self.__getFromQueue())
+                print("anomaly: queue is not empty 1")
+                sequences = np.load(self.cache_dir + np_predict_array + ".npy")
+                print("anomaly: queue is not empty 2")
+                os.remove(self.cache_dir + np_predict_array + ".npy")
+                print("anomaly: queue is not empty 3")
+                sequences_shape = sequences.shape[0]
+                print("anomaly: queue is not empty 4")
+                reconstructed_sequences = self.model.predict(sequences, batch_size=self.batch_size)
+                print("anomaly: queue is not empty 5")
+                sequences_reconstruction_cost = np.array(
+                    [np.linalg.norm(np.subtract(sequences[i], reconstructed_sequences[i])) for i in
+                    range(0, sequences_shape)])
+                print("anomaly: queue is not empty 6")
+                sa: np.ndarray = (sequences_reconstruction_cost - np.min(sequences_reconstruction_cost)) / np.max(
+                    sequences_reconstruction_cost)
+                print("anomaly: queue is not empty 7")
+                srs.append((1.0 - sa, video_counter))
+                print("anomaly: queue is not empty 8")
+                self.done_tasks[0] += 1
+                print("anomaly: queue is not empty 9")
+                # plot the regularity scores
+                # plt.plot(1.0 - sa)
+                # plt.ylabel('regularity score Sr(t)')
+                # plt.xlabel('frame t')
+                # plt.show()
+                video_counter += 1
+
+        print("anomaly: finished")
         return srs
 
     def detect_anomaly(self, video):
+        print("anomaly: detect_anomaly before get_anomaly")
         srs: list = self.get_anomaly(video)
+        print("anomaly: detect_anomaly after get_anomaly")
         total_clip_sr, video_idx = srs.pop(0)
         anomal = []
         anomal2 = []
@@ -127,18 +163,19 @@ class AnomalyDetection:
                     # anomal.append((cnt, start,(ll[x - 1][0] + 10)))
                     if cnt > self.tolerance_frames:
                         min_val = min(list(total_clip_sr)[start:(ll[x - 1][0] + 10)])
-                        anomal2.append((cnt, list(total_clip_sr).index(min_val), min_val))
+                        anomal2.append((cnt, list(total_clip_sr).index(min_val) / self.total_frames, min_val))
                     start = (ll[x][0] + 1)
                     cnt = 1
             # anomal.append((cnt, start,(ll[x - 1][0] + 10)))
             if cnt > self.tolerance_frames:
                 min_val = min(list(total_clip_sr)[start:len(total_clip_sr)])
-                anomal2.append((cnt, list(total_clip_sr).index(min_val), min_val))
+                anomal2.append((cnt, list(total_clip_sr).index(min_val) / self.total_frames, min_val))
 
         # print(anomal2)
         anomaly = {
             "anomaly": anomal2,
         }
+        print("anomaly: " + str(anomaly))
         return anomaly
 
 #
